Return unmodified version of image if Jetpack is in dev mode.
This is a lightly modified version of https://github.com/Automattic/wpcom-thumbnail-editor/pull/1

On dev sites, Jetpack is often active but Photon will not work because the content files are not accessible to the public internet. For example, VIP Quickstart dev and staging sites.
Right now, a broken image is displayed when this plugin is active and a thumbnail has been edited. This fix will at least allow the unmodified image to be displayed.